Made a New Master Script: modular this time
Allows the user to initiate the connection with the aurora NDI device only once. After this initial connection, the user can toggle magnetic tracking on and off at will. This is useful because it means you can initaite connection and not wait too long for it, while also allowing us to take segments of data without long, awkward intermissions.